tictactoe-backend/requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.4
django-cors-headers==3.7.0
pytz==2021.1
sqlparse==0.4.1
jsonschema==3.2.0
requests==2.25.1
gunicorn==20.0.4


